updated the event handler
